backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
python-dotenv==1.0.0
httpx==0.25.2
qdrant-client==1.6.8
openai==1.3.7
python-magic==0.4.27
azure-identity==1.17.0
azure-storage-blob==12.23.1
msal==1.29.0
email-validator==2.2.0
